[WIP] Add OrderSerializer so orders can be sent to Klarna
The serializer needs some more work. It woll probably break with
adjustments and maybe even certain quantities. The tax rate calculation
is a bit wonky but it seems that this is all we can get from Solidus at
the moment.
Another problem is that Klarna expects a single line item price with
taxes. Solidus appearantly does not provide that information.