Make the checkout compatible with multiple stores
It is now possible to have multiple

1. stores
2. Klarna Credit payment methods (even more than one per store)

This ensures compatibility with the solidus_multi_domain gem that is a
standard solution for having multiple frontends/shops with a single
backend/database. Payment methods ids are explicitly provided so there
is no more confusion when multiple Klarna Credit methods exists. Also
the merchant_urls are now generated using the url helpers of the current
store. That way the callback will always reach the correct store (should
be irrelevant for now as order-ids are unique across stores) and use
custom url formats.