Add the product and image urls to line items
This allows Klarna to link them in customer correspondence. We're using
the asset host for the images and when that's not available fall back to
`Spree::Store.current.url` which should work for most cases.

For the product url we can rely on the normal product_url helper from
Spree.

See #8503