Update the shipping information on shipping
Previously the shipment information was send to the Klarna API when the
order was completed/captured. The problem is that in the usual workflow
the shipment tracking information is only added when the payment was
already captured.
Instead the shipping-info is now sent when the order's shipment state
switches from ready to shipped. That way a merchant can enter tracking
information which can be sent to Klarna's API.