Remove "brand" and add feature spec
Removes the `brand` from the payment method and `supports?` from the
provider_class (the active merchant gateway subclass). While removing it
with undef_method is kind of hacky, it shouldn't be in the AM gateway in
the first place.

Also added a feature spec that goes from putting a product in the cart
to capturing the Klarna payment in the backend.