Updated Specs and Line Item Serializer to work with Active Storage
Utilization of Active Storage required a few changes with the most
consequential being in the `line_item_seriailizer`. The original serializer
built a URL out of the expected Paperclip generated url which does not work
the same way when using Active Storage. Implemented conditional which
checks if the item's url already contains "http" prior to running the
old logic or just to return the current item url. Additionally, to make
all the specs work for both Paperclip and Active Storage, `host` had to
be standardized to "http://www.example.com".