Update the place order with customer token service
Allow to place an order using the customer token without
the payment source.
This is needed since if the user already has the customer token set,
we should have the ability to
place an order without asking for the authorization again.

If the service is called with a payment source means that it is called
from the gateway and we should fetch the customer
token using the Klarna APIs.
If the payment source is not passed
means that the customer token should be retrieved from the database.