Refactor CreateCustomerTokenService to use email and address object instead of order object.
This commit changes the required arguments for the CreateCustomerTokenService class.
The required arguments before were order, authorization_token, region.
The required arguments now are email, address, authorization_token, currency and region.

This change was made due to requirements of the CreateCustomerTokenService class which has now been isolated.
This service needs to be used in places where there may not be an order to setup a subscription. In those cases we need to still create the request body and generate the customer_token. Hence, the changes in this commit to remove the dependency from the order and instead just use the required information to generate the customer_token from Klarna.