Fix issue on tax rate calculation
This commit aims to fix the issue in line_items_serializer when Spree::TaxRate is not present.
Currently the tax rate is calculated by using the value tax.source.amount from adjustments.
This causes an issue when Spree::TaxRate is not present in an app.

Here, a change has been made to calculate tax rate by using the value tax.amount of an adjustment instead to ensure that the extension works with apps where Spree::TaxRate is not present