Change base for SolidusKlarnaPayments::Api::SessionsController
This commit changes the parent controller for SolidusKlarnaPayments::Api::SessionsController from Spree::BaseCOntroller to Spree::Api::BaseController.
This is done because methods in this controller will be requested by frontend frameworks and ajax calls and this change will allow this controller to use processes defined in Solidus for Api endpoints.