Force Update of klarna_client_token
Some payments with Klarna are failing because users start a payment process, then switch
to another payment method, and try to come back. However klarna_client_token is set
to the original one so the payment fails due to invalid token.

With this PR we're trying to force the update of the klarna_client_token each time the
user changes payments or refreshes the page.