Fix Issue with Session Creation
This commit aims to fix an issue with session creation for klarna which causes customer token flow to fail.

The order_params being sent to Klarna during the session creation step have been modified to not include the billing_address now.

This change has been done in order to allow the customer_token flow for Klarna to complete successfully.
Klarna has an issue where the customer token is not generated when billing_address is passed during the session creation step, hence this has now been removed.

This however does not effect any other flow as we are passing the billing_address while placing the order.