Fix Issue when placing order with Authorization Token
There was an issue which came about due to a previous commit where session create API call was updated to get it to work for the customer token flow.

The issue was that the billing_address field was removed from the session create API call to allow it to work with the customer token flow but Klarna is very strict while checking for fields in the session create and place order API calls.
Since the billing_address field still existed in the order create API call, Klarna Order creation failed while using the authorization token to place an order.

This has now been fixed by removing the billing_address field from the Order create API call and making the fields sent to both API calls consistent.